pyfunc1/requirements.txt

attrs==23.1.0
jsonschema==4.17.3
memphis-py==1.0.5
pyrsistent==0.19.3
